+++ matrix/src/add.h
@@ -23,7 +23,7 @@
 
 #include "vect.h"
 #include "sparseVect.h"
-#include "typedefs.h"
+
 
 namespace math{
 namespace vops{
@@ -40,10 +40,10 @@ class Add
 {
 
 public:
-  void operator()( SparseVect<T> &X, SparseVect<T> &Y, SparseVect<T> &Z, ULong start = 0 ) //tested
+  void operator()( SparseVect<T> &X, SparseVect<T> &Y, SparseVect<T> &Z, unsigned long long start = 0 ) //tested
   {
-    ULong i = 0;
-    ULong j = 0;
+    unsigned long long i = 0;
+    unsigned long long j = 0;
     if ( start != 0 )
     {
       i = X.UpperBoundIdx(start);
@@ -77,12 +77,12 @@ public:
     }
   }
 
-  void operator()( ULong end, SparseVect<T> &X, SparseVect<T> &Y, SparseVect<T> &Z ) //tested
+  void operator()( unsigned long long end, SparseVect<T> &X, SparseVect<T> &Y, SparseVect<T> &Z ) //tested
   {
-    ULong i = 0;
-    ULong j = 0;
-    ULong iend = X.NNZ();
-    ULong jend = Y.NNZ();
+    unsigned long long i = 0;
+    unsigned long long j = 0;
+    unsigned long long iend = X.NNZ();
+    unsigned long long jend = Y.NNZ();
     if ( X.m_pos[iend - 1] >= end )
     {
       iend = X.UpperBoundIdx(end);
@@ -122,14 +122,14 @@ public:
     }
   }
   
-  void operator()( Vect<T> &X, SparseVect<T> &Y, Vect<T> &Z, ULong start = 0  ) //tested
+  void operator()( Vect<T> &X, SparseVect<T> &Y, Vect<T> &Z, unsigned long long start = 0  ) //tested
   {
-    ULong j = 0;
+    unsigned long long j = 0;
     if ( start != 0 )
     {
       j = Y.UpperBoundIdx(start);
     }
-    for( ULong i = start; i < Z.Size(); ++i )
+    for( unsigned long long i = start; i < Z.Size(); ++i )
     {
       if ( i == Y.m_pos[j] )
       {
@@ -146,21 +146,21 @@ public:
     }
   }
   
-  void operator()( SparseVect<T> &X, Vect<T> &Y, Vect<T> &Z, ULong start = 0  ) //tested
+  void operator()( SparseVect<T> &X, Vect<T> &Y, Vect<T> &Z, unsigned long long start = 0  ) //tested
   {
     operator()( Y, X, Z, start );
   }
   
-  void operator()( ULong end, Vect<T> &X, SparseVect<T> &Y, Vect<T> &Z ) //tested
+  void operator()( unsigned long long end, Vect<T> &X, SparseVect<T> &Y, Vect<T> &Z ) //tested
   {
     //end = Z.Size() < end ? Z.Size() : end;
-    ULong j = 0;
-    ULong jend = Y.NNZ();
+    unsigned long long j = 0;
+    unsigned long long jend = Y.NNZ();
     if ( Y.m_pos[jend - 1] >= end )
     {
       jend = Y.UpperBoundIdx(end);
     }
-    for( ULong i = 0; i < end; ++i )
+    for( unsigned long long i = 0; i < end; ++i )
     {
       if ( i == Y.m_pos[j] )
       {
@@ -177,23 +177,23 @@ public:
     }
   }
 
-  void operator()( ULong end, SparseVect<T> &X, Vect<T> &Y, Vect<T> &Z  ) //tested
+  void operator()( unsigned long long end, SparseVect<T> &X, Vect<T> &Y, Vect<T> &Z  ) //tested
   {
     operator()( end, Y, X, Z );
   }
   
-  void operator()( Vect<T> &X, Vect<T> &Y, Vect<T> &Z, ULong start = 0  ) //tested
+  void operator()( Vect<T> &X, Vect<T> &Y, Vect<T> &Z, unsigned long long start = 0  ) //tested
   {
-    for ( ULong i = start; i < X.Size(); ++i )
+    for ( unsigned long long i = start; i < X.Size(); ++i )
     {
       Z[i] = X[i] + Y[i];
     }
   }
 
-  void operator()( ULong end, Vect<T> &X, Vect<T> &Y, Vect<T> &Z  ) //tested
+  void operator()( unsigned long long end, Vect<T> &X, Vect<T> &Y, Vect<T> &Z  ) //tested
   {
     //end = Z.Size() < end ? Z.Size() : end;
-    for ( ULong i = 0; i < end; ++i )
+    for ( unsigned long long i = 0; i < end; ++i )
     {
       Z[i] = X[i] + Y[i];
     }
